extra_install/asterisk_connector/requirements.txt
pymustache
ari
pudb
requests
paho-mqtt
